Configure Network Settings for Event Hub with Managed Identity (#401)
This commit involves changes to the Event Hub and its associated components. The purpose of these modifications is to enable the Event Hub to function with the "All Networks" network setting, and use IAM and Managed Identities to keep the Event Hub secure. 

To achieve this, the recommended approach was to implement Managed Identity for both the sender (Event Grid) and the receiver (Azure Function). This facilitates secure access to both the Event Hub and the Event Hub Namespace, aligning with the prescribed architectural guidelines.