Remove Dh2HealthCheck certs and tf files (#1017)
If applied, this PR removes no longer relevant 
certificates and terraform files for those environments 
where the DH2 health check is no longer relevant. 

This is done as Migration is switching to the 
correct Time Series Synchronization setup in DH2. 
Making some of the environments obsolete for
this DH2 health check.

The environment where the DH2 health check is 
still relevant are untouched.